Enable strings category methods tests
